move components one level up
