+++ ui/src/pages/namespace/Gateway/Routes/components/Methods.tsx
