+++ ui/src/pages/namespace/Gateway/Routes/components/PublicPath.tsx
